Separation `Comment` and `CommentForm`
