+++ app/views/kitten/components/form/comment-form.html.erb
@@ -1,0 +1,25 @@
+<div class="k-Grid">
+  <div class="k-Grid__col k-Grid__col--10">
+    <%= example 'Comment' do %>
+      <%= react_component 'CommentForm', props: {
+        ownerImgProps: { src: 'https://placehold.it/80x80/caf4fe/caf4fe' },
+        ownerName: 'Camille',
+        ownerTime: '2 min',
+        placeholder: 'Laisser un commentaire sur le projet',
+      } %>
+    <% end %>
+  </div>
+
+  <div class="k-Grid__col k-Grid__col--10">
+    <%= example 'Comment' do %>
+      <%= react_component 'CommentForm', props: {
+        ownerImgProps: { src: 'https://placehold.it/80x80/caf4fe/caf4fe' },
+        ownerName: 'Camille',
+        ownerTime: '2 min',
+        placeholder: 'Laisser un commentaire sur le projet',
+        commentButton: 'Commenter',
+        cancelButton: 'Annuler',
+      } %>
+    <% end %>
+  </div>
+</div>
