+++ app/views/kitten/components/form/comment.html.erb
@@ -1,15 +1,1 @@
-<div class="k-Grid">
-  <div class="k-Grid__col k-Grid__col--10">
-    <%= example 'Comment' do %>
-      <%= react_component 'Comment', props: {
-        ownerImgProps: { src: 'https://placehold.it/80x80/caf4fe/caf4fe' },
-        ownerName: 'Camille',
-        ownerTime: '2 min',
-        placeholder: 'Laisser un commentaire sur le projet',
-        commentButton: 'Commenter',
-        cancelButton: 'Annuler',
-        isValid: true,
-      } %>
-    <% end %>
-  </div>
-</div>
+comment.html.erb
