+++ spec/dummy/client/entries/app-kitten.js
@@ -82,7 +82,7 @@ import { KarlFooterKisskiss } from 'kitten/karl/footer/footer-kisskiss'
 // Form
 import { Checkbox } from 'kitten/components/form/checkbox'
 import { CheckboxSet } from 'kitten/components/form/checkbox-set'
-import { Comment } from 'kitten/components/form/comment'
+import { CommentForm } from 'kitten/components/form/comment-form'
 import { FormAmountAndCurrency }
   from 'kitten/components/form/form-amount-and-currency'
 import { FormPhoneNumber } from 'kitten/components/form/form-phone-number'
@@ -389,7 +389,7 @@ ReactOnRails.register({
   // Form
   Checkbox,
   CheckboxSet,
-  Comment,
+  CommentForm,
   FormAmountAndCurrency,
   FormPhoneNumber,
   Label,
